+++ src/settings.ml
@@ -149,6 +149,14 @@ let server_syntaxDocumentation_setting =
     ~to_json:Jsonoo.Encode.bool
 ;;
 
+let server_typedHolesConstructAfterNavigate_setting =
+  create_setting
+    ~scope:ConfigurationTarget.Workspace
+    ~key:"ocaml.commands.typedHoles.constructAfterNavigate"
+    ~of_json:Jsonoo.Decode.bool
+    ~to_json:Jsonoo.Encode.bool
+;;
+
 let server_constructRecursiveCalls_setting =
   create_setting
     ~scope:ConfigurationTarget.Workspace